+++ Sources/Sheeeeeeeeet/Menu/Items/CollectionItem.swift
@@ -12,10 +12,19 @@ import Foundation
  Collection items can be used to present item collections of
  a certain item type, e.g. in a collection view.
  
- TODO: Implement `ActionSheetItemConvertible`, by decoupling
- the item type from the cell concept.
+ For now `CollectionItemType` has many presentation-specific
+ properties, which is legacy reasons, where collection items
+ were always collection view cells. This will be solved in a
+ future version.
+ 
+ TODO: For now, creating an `ActionSheet` with a `Menu` that
+ contains a `CollectionItem` provides no way of updating the
+ action sheet from within the `itemSelectionAction`. You can
+ thus only use this if you do not need to update your action
+ sheet when items are selected.
+ 
  */
-open class CollectionItem<T>: MenuItem {
+open class CollectionItem<T: CollectionItemType>: MenuItem {
     
     
     // MARK: - Initialization
@@ -35,7 +44,7 @@ open class CollectionItem<T>: MenuItem {
     
     // MARK: - Typealiases
     
-    public typealias ItemAction = (_ cell: T, _ index: Int) -> Void
+    public typealias ItemAction = (_ item: T, _ index: Int) -> Void
     
     
     // MARK: - Properties
@@ -49,7 +58,27 @@ open class CollectionItem<T>: MenuItem {
     // MARK: - ActionSheetItemConvertible
     
     override func toActionSheetItem() -> ActionSheetItem {
-        assertionFailure("This is not yet supported.")
-        return super.toActionSheetItem()
+        ActionSheetCollectionItem(
+            itemCellType: itemType,
+            itemCount: itemCount,
+            setupAction: itemSetupAction,
+            selectionAction: itemSelectionAction
+        )
     }
+}
+
+
+/**
+ This protocol must be implemented by any item that is to be
+ used together with an `CollectionItem` and presented in the
+ collection, regardless of what that collection is.
+ */
+public protocol CollectionItemType {
+    
+    static var defaultSize: CGSize { get }
+    static var leftInset: CGFloat { get }
+    static var rightInset: CGFloat { get }
+    static var topInset: CGFloat { get }
+    static var bottomInset: CGFloat { get }
+    static var itemSpacing: CGFloat { get }
 }